ダッシュボードのページにアプリの説明を記入

+++ app/Http/Controllers/HomeController.php
@@ -95,8 +95,10 @@ class HomeController extends Controller
      */
     public function destroy($id)
     {
-        $friend = User::find($id);
-        $friend->delete();
+        if($id != 1){
+            $friend = User::find($id);
+            $friend->delete();
+        }
 
         return to_route('home.index');
     }
